fix(lsp): move pull diagnostics out of feature dispatch
